Refactor baseURL imports in multiple components: Updated AdminSideBar, AvailabilityManager, AvailabilityViewer, PartnerSideBar, and TourDetail components to use a consistent import path for baseURL, enhancing code organization and maintainability.
